+++ python/lsst/ts/wep/utils/modelUtils.py
@@ -43,8 +43,8 @@ def forwardModelPair(
     seeing: Union[float, None] = None,
     skyLevel: Union[float, None] = None,
     bandLabel: str = "r",
-    fieldAngleIntra: tuple = (0.23, -0.6),
-    fieldAngleExtra: tuple = (0.23, -0.6),
+    fieldAngleIntra: Union[tuple, None] = None,
+    fieldAngleExtra: Union[tuple, None] = None,
     blendOffsetsIntra: Union[np.ndarray, tuple, list, None] = None,
     blendOffsetsExtra: Union[np.ndarray, tuple, list, None] = None,
     instConfig: Union[str, dict, Instrument] = "policy:instruments/LsstCam.yaml",
@@ -88,11 +88,15 @@ def forwardModelPair(
         The name of the band to simulate donuts in.
         (the default is "r")
     fieldAngleIntra : tuple, optional
-        Field angle, in degrees, of the intrafocal donut.
-        (the default is (0.23, -0.6))
+        Field angle, in degrees, of the intrafocal donut. If the angle is
+        only specified for the intra or extrafocal donut, both donuts use
+        that angle. If neither is specified, the same random angle is used
+        for both. (the default is None)
     fieldAngleExtra : tuple, optional
-        Field angle, in degrees, of the extrafocal donut.
-        (the default is (0.23, -0.6))
+        Field angle, in degrees, of the extrafocal donut. If the angle is
+        only specified for the intra or extrafocal donut, both donuts use
+        that angle. If neither is specified, the same random angle is used
+        for both. (the default is None)
     blendOffsetsIntra : Iterable or None, optional
         The blend offsets of the intrafocal donut.
         (the default is None)
@@ -117,11 +121,13 @@ def forwardModelPair(
     # Create the ImageMapper that will forward model the images
     mapper = ImageMapper(instConfig=instConfig)
 
+    # And the random number generators
+    rng = np.random.default_rng(seed)
+
     # Generate random Zernikes?
-    if zkCoeff is None:
-        rng = np.random.default_rng(seed)
-        zkCoeff = rng.normal(0, zkNorm / np.arange(1, jmax - 2) ** 1.5, size=jmax - 3)
-        zkCoeff = np.clip(zkCoeff, -zkMax, +zkMax)
+    _zkCoeff = rng.normal(0, zkNorm / np.arange(1, jmax - 2) ** 1.5, size=jmax - 3)
+    _zkCoeff = np.clip(_zkCoeff, -zkMax, +zkMax)
+    zkCoeff = _zkCoeff if zkCoeff is None else zkCoeff
 
     # Sample random seeing and skyLevel?
     _seeing = rng.uniform(0.3, 1.5)
@@ -129,13 +135,35 @@ def forwardModelPair(
     _skyLevel = 10 ** rng.uniform(2, 9)
     skyLevel = _skyLevel if skyLevel is None else skyLevel
 
+    # Sample random field angles?
+    _maxAngle = np.max(
+        [
+            params["thetaMax"]
+            for edge in mapper.instrument.maskParams.values()
+            for params in edge.values()
+        ]
+    )
+    _fieldAngleRadius = rng.uniform(0, _maxAngle)
+    _fieldAngleAzimuth = rng.uniform(0, 2 * np.pi)
+    _fieldAngle = _fieldAngleRadius * np.array(
+        [np.cos(_fieldAngleAzimuth), np.sin(_fieldAngleAzimuth)]
+    )
+    if fieldAngleIntra is None and fieldAngleExtra is None:
+        fieldAngleIntra = tuple(_fieldAngle)
+        fieldAngleExtra = tuple(_fieldAngle)
+    elif fieldAngleIntra is None:
+        fieldAngleIntra = fieldAngleExtra
+    elif fieldAngleExtra is None:
+        fieldAngleExtra = fieldAngleIntra
+
     # Create a Kolmogorov kernel
-    atm = galsim.Kolmogorov(fwhm=seeing)
-    atmKernel = atm.drawImage(
-        nx=nPix // 2,
-        ny=nPix // 2,
-        scale=mapper.instrument.pixelScale,
-    ).array
+    if seeing > 0:
+        atm = galsim.Kolmogorov(fwhm=seeing)
+        atmKernel = atm.drawImage(
+            nx=nPix // 2,
+            ny=nPix // 2,
+            scale=mapper.instrument.pixelScale,
+        ).array
 
     # Calculate background per pixel level from skyLevel
     background = skyLevel * mapper.instrument.pixelScale**2
@@ -155,16 +183,19 @@ def forwardModelPair(
     if blendOffsetsIntra is None:
         nBlends = 0
     else:
-        nBlends = len(blendOffsetsIntra)
+        offsets = np.atleast_2d(blendOffsetsIntra)
+        nBlends = len(offsets)
         centralDonut = intraStamp.image.copy()
-        for blendShift in blendOffsetsIntra:
+        for blendShift in offsets:
             intraStamp.image += shift(centralDonut, blendShift[::-1])
     # Convolve with Kolmogorov atmosphere
-    intraStamp.image = convolve(intraStamp.image, atmKernel, mode="same")
+    if seeing > 0:
+        intraStamp.image = convolve(intraStamp.image, atmKernel, mode="same")
     # Normalize the flux
     intraStamp.image *= fluxIntra * (1 + nBlends) / intraStamp.image.sum()
     # Poisson noise
     intraStamp.image += background
+    intraStamp.image = np.clip(intraStamp.image, 0, None)
     intraStamp.image = rng.poisson(intraStamp.image).astype(float)
     intraStamp.image -= background
 
@@ -183,16 +214,19 @@ def forwardModelPair(
     if blendOffsetsExtra is None:
         nBlends = 0
     else:
-        nBlends = len(blendOffsetsExtra)
+        offsets = np.atleast_2d(blendOffsetsExtra)
+        nBlends = len(offsets)
         centralDonut = extraStamp.image.copy()
-        for blendShift in blendOffsetsExtra:
+        for blendShift in offsets:
             extraStamp.image += shift(centralDonut, blendShift[::-1])
     # Convolve with Kolmogorov atmosphere
-    extraStamp.image = convolve(extraStamp.image, atmKernel, mode="same")
+    if seeing > 0:
+        extraStamp.image = convolve(extraStamp.image, atmKernel, mode="same")
     # Normalize the flux
     extraStamp.image *= fluxExtra / extraStamp.image.sum()
     # Poisson noise
     extraStamp.image += background
+    extraStamp.image = np.clip(extraStamp.image, 0, None)
     extraStamp.image = rng.poisson(extraStamp.image).astype(float)
     extraStamp.image -= background